# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a kingford backend  server Petstore server.",
        "title": "Swagger kingford backend API",
        "contact": {
            "name": "API Support",
            "email": "stevekingford@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/api/collection": {
            "get": {
                "description": "获取分页列表",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "创建信息",
                "parameters": [
                    {
                        "description": "创建信息",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collection.CreateService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            }
        },
        "/api/collection-category": {
            "get": {
                "description": "获取分页列表",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CollectionItem"
                ],
                "summary": "分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CollectionItem"
                ],
                "summary": "创建信息",
                "parameters": [
                    {
                        "description": "创建信息",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.CreateItemService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            }
        },
        "/api/collection-category/{id}": {
            "get": {
                "description": "通过id获取信息",
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "CollectionItem"
                ],
                "summary": "通过Id获取信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "通过id修改信息",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CollectionItem"
                ],
                "summary": "通过id修改信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改信息",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.UpdateItemService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "通过id删除信息",
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "CollectionItem"
                ],
                "summary": "通过id删除信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            }
        },
        "/api/collection/{id}": {
            "get": {
                "description": "通过id获取信息",
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "通过Id获取信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "通过id修改信息",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "通过id修改信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改信息",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collection.UpdateService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "通过id删除信息",
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "通过id删除信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/global.Response"
                        }
                    }
                }
            }
        },
        "/system/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "用户登录",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "用户注册",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/user": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取用户列表",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/user/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取用户列表",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "修改用户信息",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "删除用户信息",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "category.CreateItemService": {
            "type": "object",
            "properties": {
                "collectionId": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "category.UpdateItemService": {
            "type": "object",
            "properties": {
                "collectionId": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "collection.CreateService": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "collection.UpdateService": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "global.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}